remove(rpc): we decided to remove rpc and use core.Client directly; testutils: move Core mock to core package
